Started up on a generic linked list implementation
Trying to make this generic for the purposes of potiential reuse in the
future. For this reason, the idea of having a toolkit repo that I can
pull from in future projects was added to the TODOs.txt.

For this project, the LinkedList will be used as the data structure for
the BuildSequence list. The idea is that if this is a linked list we can
allow plugins to insert new build steps into the list.