Makefile now using wildcards appropriately.
Should not need to modify the makefile anymore when files are added.
Currently compiling all of the toolkit source files into .o files inside
the lib directory. Then compiling the GregBuild source files with the
object files from this lib folder. This way make should skip compiling
the toolkit lib if it has not changed. Also currently excluding the
LinkedList.c since we have a test main function in there. When that work
is finished and that main function is removed, it can be included again.

In the future we may need to modify this to ensure the GregBuild
executable only includes the lib/*.o files that it needs and not the
entire toolkit.