Honing in on the correct values for the TestMain.c
and TestMain.h files. Still need to look at verifying the per test case
values. Should also print a different message if no test cases are found
instead of "All Test Passed Successfully".

Finally, the were some make file improvements so that copying the
GregTest.dll step is omitted if we already have it. Still need a better
solution for only building the toolkit when needed and only linking as
much from the toolkit as is needed. Perhaps a GregCToolkit.dll is a way
to go with this for the first step in the future.