Added functionName member to BuildSequenceStep
This allows us to search through out buildsequence looking for a
specific step. We use this to find the index of any core build sequence  step so a plugin can
add a its step before or after the given core reference step.

Only processing plugin functions before and after the first core step at
this point.