Split PluginProcessor out from PluginLoader
Fixed changed malloc(sizeof(char)) to malloc(WINDOWS_MAX_PATH_LENGTH * sizeof(char))
throughout the repo. The resolves issues with some of the string not being long enough to hold
all of the desired contents.

Added some prints to executeBuildSequence() to verify this commit and commits in the future related to build sequence re-ordering with plugins.

Move the Plugin and PluginList struct definitions to PluginDefs.h

Corrected loadingTestAndSourceFiles to loadTestsAndSourceFiles for plugin processing