Fixed bugs in plugin loader.
The first issue was the way I was initializing the beforeFunction and
afterFunction strings. Noticed that sometimes an extra space was put at
the start and therefore was failing the strcmp checks.

Second issue was the value used by the beforeIndex. It was

int beforeIndex = index - 1;

But the way the linked list insert works, it should have been

int beforeIndex = index;

This was fixed as well. After these two changes were made, the plugin
loader worked as expected through a few test runs.