Refactored to use CommandLineOptionsList instead of linked list.
This allows for less complex memory management, and removes a noticeable
memory leak. The negative to this decision is that if we need to order
the command line options list in the future, we will have to reorder
after the fact instead of doing the ordering as we add to the list.

The only place I can think of right now that we might want to reorder the
list in the future is in the supported options printout, where we
might want them to be order alphabetically.